refactoring: aplhavantage data models moved to models package
